Implemented services and controllers for new models
